IndustryViewHolder в отдельный класс
